feat(list): Create rows and tiles
Currently using example responce as login is still broken. thumnail object seems
to be free'd by the time the thread is done downloading the image for it and
then causes a segfault. Not sure why the thumbnail is being free'd, something to
do with QMainWindow thinking its not important anymore.